requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-taggit==5.0.1
gunicorn==22.0.0
modeltranslation==0.25
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
pytz==2024.1
shortuuid==1.0.13
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
psycopg2-binary
